requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
aiosqlite
pydantic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
python-multipart
slowapi
